+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `PartialSpecificationRuleExtractor` that extracts rules from partially built
   specifications and two subclasses with specific applications of that.
 - Make specification iterable
+- Add a `get_comb_class` method to combinatorial specification
 
 ### Changed
 - If a rule in a specification cannot be sanity checked (e.g., counting is not